+++ crates/adapters/databento/src/symbology.rs
@@ -128,6 +128,12 @@ pub fn get_nautilus_instrument_id_for_record(
         (msg.hd.instrument_id, msg.ts_recv)
     } else if let Some(msg) = record.get::<dbn::InstrumentDefMsg>() {
         (msg.hd.instrument_id, msg.ts_recv)
+    } else if let Some(msg) = record.get::<dbn::Cmbp1Msg>() {
+        (msg.hd.instrument_id, msg.ts_recv)
+    } else if let Some(msg) = record.get::<dbn::CbboMsg>() {
+        (msg.hd.instrument_id, msg.ts_recv)
+    } else if let Some(msg) = record.get::<dbn::TbboMsg>() {
+        (msg.hd.instrument_id, msg.ts_recv)
     } else {
         anyhow::bail!("DBN message type is not currently supported")
     };